refactor(core): move durable loop onto AgenticLoopBuilder as a subclass
Introduce DurableAgenticLoopBuilder extending AgenticLoopBuilder: the durable
agentic workflow's composition, workflow options, and dowhile continuation
predicate move onto overrides of the shared builder shape, and
createDurableAgenticWorkflow becomes a thin wrapper. Durable step files are
untouched and the composition is moved verbatim, so this is
behavior-preserving.

The base class's overridable surface is typed with engine-agnostic LoopStep /
LoopWorkflow handles since the two engines build steps with different ids and
schemas; each engine keeps its concrete types internally. With both loops on
one class shape, the remaining divergence is visible as overrides — the
shared-core migration now shrinks these overrides method by method until only
runtime hooks (state resolution, chunk transport, engine selection) remain.
